First pass at rationalising status area code
First pass at rationalising status area code  - more work needed!
